+++ db/connection.js
@@ -3,13 +3,4 @@ const mongoose = require('mongoose')
 
 mongoose.connect(process.env.MONGOLAB_CRIMSON_URI || 'mongodb://localhost/authors')
 
-const AuthorSchema = new mongoose.Schema({
-  name: String,
-  publisher: String,
-  age: Number,
-  active: Boolean
-})
-
-const Author = mongoose.model('Author', AuthorSchema)
-
 module.exports = mongoose
