+++ db/seeds.json
@@ -1,0 +1,20 @@
+[
+  {
+    "name": "3213",
+    "publisher": "sasdsa",
+    "age": 23,
+    "active": true
+  },
+  {
+    "name": "3213",
+    "publisher": "sasdsa",
+    "age": 23,
+    "active": true
+  },
+  {
+    "name": "3213",
+    "publisher": "sasdsa",
+    "age": 23,
+    "active": true
+  }
+]
